HW1_Image-Classification-Problem/requirements.txt
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
PyYAML==6.0.2
torch==2.6.0+cu124
torchvision==0.21.0
